notification_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
python-dotenv
kafka-python
httpx
firebase-admin